Restructure file hierarchy; Add tests for snippet_id endpoint
