+++ api_tests/config.py
@@ -1,0 +1,7 @@
+import json
+from os.path import dirname, abspath
+
+def parse():
+    config_path = dirname(__file__) + '/config.json'
+    config_dict = json.load(open(config_path, 'r'))
+    return config_dict
