+++ api_tests/users-user_id-snippets-snippet_id/__init__.py
@@ -1,0 +1,1 @@
+## Establish this directory as a python module
